implemented query for astar

+++ code/src/mydijkstra/src/dijkstra_function.cpp
@@ -8,9 +8,9 @@ using namespace std;
 using namespace boost;
 typedef adjacency_list<vecS, vecS, undirectedS, Vertex,Edge> G;
 //calculates the shortest path,returns the pathsize and also the path as an array of edges
-int compute_dijkstra(int source,int target,Edge *edges,int num_vertices,int num_edges,Edge **path)
+int compute_dijkstra(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path)
 {
-	int path_size=0;
+	int64_t path_size=0;
 	typedef My_dijkstra<G> Graph;
 	Graph g(DIRECTED,num_vertices);
 	g.initialize_graph(edges,num_edges);
